Move dev_null() and diagnostics() to their respective modules, clean up imports
